change balance column from int to bigint

+++ database/migrations/2021_12_18_215349_create_wallet_table.php
@@ -16,7 +16,7 @@ class CreateWalletTable extends Migration
         Schema::create('wallets', function (Blueprint $table) {
             $table->id();
             $table->unsignedBigInteger('userId');
-            $table->integer('balance')->default(0);
+            $table->bigInteger('balance')->default(0);
             $table->string('bankId')->default(null);
             $table->timestamps();
